+++ ethsigner/main/src/test/java/tech/pegasys/ethsigner/main/CommandLineConfigTest.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.ethsigner;
+package tech.pegasys.ethsigner.main;
 
 import static org.assertj.core.api.Assertions.assertThat;
